Combine single and bulk write code paths using writev(2) scatter-gather operations

+++ utils.h
@@ -28,14 +28,13 @@
 #include <stdint.h> /* uint32_t */
 #include <pthread.h>
 
-
 char* path_concat (const char *dir, const char *file);
 
 struct inotify_event* create_inotify_event (int         wd,
                                             uint32_t    mask,
                                             uint32_t    cookie,
                                             const char *name,
-                                            int        *event_len);
+                                            size_t     *event_len);
 
 ssize_t safe_read   (int fd, void *data, size_t size);
 ssize_t safe_write  (int fd, const void *data, size_t size);
